+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 
 First version. The following spatial operations are available:
 
-- Calculate the centerline for a polygon (#1).
+- centerline: calculates the approximate centerline for a polygon (#1)
+- view_angles: calculates the angles where a geometry can be seen from a viewpoint (#7)